fix: -Wsign-compare in PrintObject, PrintObjectSlice, PrintConfig, Milling, 22 -> 0
PrintObject.cpp 6 -> 0, PrintConfig.cpp 6 -> 0, PrintObjectSlice.cpp 5 -> 0,
Milling/MillingPostProcess.cpp 5 -> 0 (that file is now free of GCC warnings
entirely). Running total 101 of 206.

One of these is a real bug, in the vertical-shell solid-fill pass:

    const int min_layer_no_solid = region.config().bottom_solid_layers.value - 1;
    ...
    if (nb_perimeter_layers_for_solid_fill != 0 && (idx_layer > min_layer_no_solid || ...

idx_layer is a size_t. bottom_solid_layers has def->min = 0, so with no bottom
solid layers min_layer_no_solid is -1, which converts to SIZE_MAX and makes the
test impossible to pass - silently disabling the whole solid-fill-over-perimeters
branch. The name and the `- 1` say the intent is "this layer is above the last
solid bottom layer", i.e. idx_layer >= bottom_solid_layers, and every layer is
above layer -1. Now compares int(idx_layer) at all three sites.

Checked in Python: int(idx_layer) > min_layer_no_solid matches that intent for
every bottom_solid_layers value, and differs from the old behaviour ONLY when the
option is 0. That difference is the bug being fixed.

The two `i - idx_layer < nb_perimeter_layers_for_solid_fill` tests move to the
signed domain too. Both loops set `int i = int(idx_layer) +/- 1` and step away
from idx_layer, so the subtraction never actually underflows; verified no
disagreement with the old form across the reachable range.

MillingPostProcess.cpp is pattern B throughout, not size_t: first_point_idx and
second_point_idx are -1 sentinels that get tested with `> -1` and `== -1`, so the
int32_t indices have to stay signed and the sizes get cast instead.
PrintObjectSlice's elephant-foot conditions compare a size_t layer_id against
first_layer_size_compensation_layers (def->min = 1) and raft_layers (def->min = 0)
- both provably non-negative, so the casts are exact. The rest are plain forward
index loops.

Verified. All four TUs compile with -Wsign-compare at 0, no other flag's count
moved. Runtime against the pre-campaign binary from eae5ede on the FrankenPrinter
profile, 2 runs compared against every available before-run: 9 of 11 panel configs
exact, including all 7 strict gates and bedcover and both thinwalls variants. pins
is structurally identical (0 differing layers) with 0.000149 mm float drift, inside
its measured range. handle is excluded as a gate - it differs 12-31 layers
run-to-run by itself.

NOT runtime-exercised: the min_layer_no_solid difference needs
bottom_solid_layers = 0, and this profile uses 4. The changed lines do execute
(solid_over_perimeters = 2, so the branch is taken) but the -1 path is not
reachable here, so that specific fix rests on the Python check and the stated
intent rather than on a measured before/after.

+++ src/libslic3r/PrintConfig.cpp
@@ -10264,7 +10264,7 @@ void PrintConfigDef::handle_legacy_composite(DynamicPrintConfig &config, std::ma
         double external_perimeter_speed = config.get_computed_value("external_perimeter_speed");
         double max = external_perimeter_speed;
         double min = external_perimeter_speed;
-        for (int x = 0; x < values.size(); ++x) {
+        for (size_t x = 0; x < values.size(); ++x) {
             if (values[x].percent) {
                 min = std::min(min, values[x].get_abs_value(external_perimeter_speed));
                 max = std::max(max, values[x].get_abs_value(external_perimeter_speed));
@@ -10287,7 +10287,7 @@ void PrintConfigDef::handle_legacy_composite(DynamicPrintConfig &config, std::ma
         opt.set_can_be_disabled();
         // extract values
         Pointfs graph_curve;
-        for (int x = 0; x < values.size(); ++x) {
+        for (size_t x = 0; x < values.size(); ++x) {
             double speed = values[x].get_abs_value(external_perimeter_speed);
             speed = std::clamp(speed, min, max);
             double percent = (speed - min) / (max - min);
@@ -10351,10 +10351,10 @@ void PrintConfigDef::handle_legacy_composite(DynamicPrintConfig &config, std::ma
             }
         }
         // while there is a value
-        for(int idx = 0 ;idx < enable_dynamic_fan_speeds.size(); ++idx) {
+        for(size_t idx = 0 ;idx < enable_dynamic_fan_speeds.size(); ++idx) {
             // extract values
             Pointfs graph_curve;
-            for (int x = 0; x < values.size(); ++x) {
+            for (size_t x = 0; x < values.size(); ++x) {
                 graph_curve.push_back(Vec2d(x*OVERLAP_GRAPH_STEP_PCT, values[x].get_at(idx)));
             }
             if (external_perimeter_fan_speed && external_perimeter_fan_speed->is_enabled(idx)) {
@@ -10370,7 +10370,7 @@ void PrintConfigDef::handle_legacy_composite(DynamicPrintConfig &config, std::ma
         }
         // recreate fan speed graph
         opt.set(graph_data);
-        for (int idx = 0; idx < enable_dynamic_fan_speeds.size(); ++idx) {
+        for (size_t idx = 0; idx < enable_dynamic_fan_speeds.size(); ++idx) {
             opt.set_enabled(enable_dynamic_fan_speeds.get_at(idx), idx);
         }
         config.set_key_value("overhangs_dynamic_fan_speed", opt.clone());
@@ -11433,7 +11433,7 @@ std::map<std::string, std::string> PrintConfigDef::to_prusa(t_config_option_key&
             const ConfigOptionFloats *nozzle_diameters = all_conf.option<ConfigOptionFloats>(KEY_NOZZLE_DIAMETER);
             assert(current_opt && nozzle_diameters);
             assert(current_opt->size() == nozzle_diameters->size());
-            for (int i = 0; i < current_opt->size(); i++) {
+            for (size_t i = 0; i < current_opt->size(); i++) {
                 computed_opt.set_at(current_opt->get_abs_value(i, nozzle_diameters->get_at(i)), i);
             }
             assert(computed_opt.size() == nozzle_diameters->size());
